Pre-dispatch all for information for the preprocessing.
There might be a way around this but for now this is the simplest
solution. The problem is that sqlite is not thread safe and
the iterator for the parallel map from joblib are executed
in different processes. The problem rarely manifests, but if it
does it is not nice. This commit should avoid it for good.